requirements.txt
pandas==2.1.0
google-generativeai==0.3.1
matplotlib==3.8.0
seaborn==0.13.0
click==8.1.7
python-dotenv==1.0.0
better-profanity==0.7.0 